Changed IfcGeometryModelRenderer to a general GeometryModelRenderer a child class for IfcModels and a child class for OffModels
