+++ UnitTests/Utilities/IfcGeometryModelRenderer/src/IfcGeometryModelRenderer.cpp
@@ -1,5 +1,5 @@
 /*
-    Copyright (c) 2020 Technical University of Munich
+    Copyright (c) 2021 Technical University of Munich
     Chair of Computational Modeling and Simulation.
 
     TUM Open Infra Platform is free software; you can redistribute it and/or modify
@@ -17,59 +17,10 @@
 
 #include "IfcGeometryModelRenderer.h"
 
-struct WorldBuffer
-{
-    buw::Matrix44f viewProjection;
-    buw::Matrix44f projection;
-    buw::Matrix44f view;
-    buw::Vector3f cam;
-    buw::Matrix44f rotation;
-};
 
 IfcGeometryModelRenderer::IfcGeometryModelRenderer(const buw::ReferenceCounted<buw::IRenderSystem>& renderSystem)
-    : renderSystem_{renderSystem}
+    : GeometryModelRenderer(renderSystem)
 {
-    backBuffer_ = renderSystem_->getBackBufferTarget();
-    backBuffer_->makeCPUReadable();
-
-    width_ = backBuffer_->width();
-    height_ = backBuffer_->height();
-
-    camera_ = buw::makeReferenceCounted<buw::Camera>();
-    cameraController_ = buw::makeReferenceCounted<buw::CameraController>(camera_);
-
-    camera_->frustum() = buw::CameraFrustum(width_, height_, 0.5f, 50000.f, buw::eProjectionType::Perspective);
-    camera_->transformation().offset() = 15;
-    camera_->transformation().yaw() = buw::constantsf::pi_over_4();
-
-    buw::texture2DDescription dsvTD;
-    dsvTD.width = width_;
-    dsvTD.height = height_;
-    dsvTD.format = buw::eTextureFormat::D24_UnsignedNormalizedInt_S8_UnsignedInt;
-    dsvTD.data = nullptr;
-    dsvTD.isCpuReadable = false;
-    dsvTD.useMSAA = renderSystem_->getMSAAEnabled();
-    depthStencilMSAA_ = renderSystem_->createTexture2D(dsvTD, buw::eTextureBindType::DSV);
-
-    viewport_ = renderSystem_->createViewport(buw::viewportDescription(width_, height_));
-
-    buw::constantBufferDescription cbd0;
-    cbd0.sizeInBytes = sizeof(WorldBuffer);
-    cbd0.data = nullptr;
-    worldBuffer_ = renderSystem_->createConstantBuffer(cbd0);
-
-    WorldBuffer world;
-    world.viewProjection = camera_->viewProjectionMatrix();
-    world.projection = camera_->frustum().projectionMatrix();
-    world.view = camera_->transformation().viewMatrix();
-    world.cam = (camera_->transformation().transformationMatrix() * buw::Vector4f(0, 0, 0, 1)).block<3, 1>(0, 0);
-    world.rotation = camera_->transformation().rotationMatrix();
-
-    buw::constantBufferDescription cbd;
-    cbd.sizeInBytes = sizeof(WorldBuffer);
-    cbd.data = &world;
-    worldBuffer_->uploadData(cbd);
-
     ifcGeometryEffect_ = buw::makeReferenceCounted<oip::IfcGeometryEffect>(
         renderSystem_.get(), viewport_, depthStencilMSAA_, worldBuffer_);
     ifcGeometryEffect_->init();
@@ -79,13 +30,6 @@ IfcGeometryModelRenderer::~IfcGeometryModelRenderer()
 {
     model_.reset();
     ifcGeometryEffect_.reset();
-    worldBuffer_.reset();
-    viewport_.reset();
-    depthStencilMSAA_.reset();
-    backBuffer_.reset();
-    renderSystem_.reset();
-    cameraController_.reset();
-    camera_.reset();
 }
 
 void IfcGeometryModelRenderer::fitViewToModel() const
@@ -97,8 +41,7 @@ void IfcGeometryModelRenderer::fitViewToModel() const
     camera_->tick(1.0f);
 }
 
-void IfcGeometryModelRenderer::setModel(
-    const std::shared_ptr<oip::IfcModel>& model)
+void IfcGeometryModelRenderer::setModel(const std::shared_ptr<oip::IfcModel>& model)
 {
     model_.reset();
     model_ = model;
@@ -109,56 +52,10 @@ void IfcGeometryModelRenderer::setModel(
     fitViewToModel();
 }
 
-void IfcGeometryModelRenderer::clearBackBuffer()
-{
-    float color[4] = {0.0f, 0.0f, 0.0f, 0.0f};
-    renderSystem_->clearRenderTarget(backBuffer_, color);
-    renderSystem_->clearDepthStencilView(depthStencilMSAA_);
-}
-
 void IfcGeometryModelRenderer::repaint()
 {
     clearBackBuffer();
     updateWorldBuffer();
     ifcGeometryEffect_->render();
     renderSystem_->present();
-}
-
-void IfcGeometryModelRenderer::updateWorldBuffer()
-{
-    WorldBuffer world;
-    world.viewProjection = camera_->viewProjectionMatrix();
-    world.projection = camera_->frustum().projectionMatrix();
-    world.view = camera_->transformation().viewMatrix();
-    world.cam = (camera_->transformation().transformationMatrix() * buw::Vector4f(0, 0, 0, 1)).block<3, 1>(0, 0);
-    world.rotation = camera_->transformation().rotationMatrix();
-
-    buw::constantBufferDescription cbd;
-    cbd.sizeInBytes = sizeof(WorldBuffer);
-    cbd.data = &world;
-    worldBuffer_->uploadData(cbd);
-}
-
-buw::Image4b IfcGeometryModelRenderer::getBackBufferImage() const
-{
-    auto backBufferImage = buw::Image4b(width_, height_);
-    renderSystem_->downloadTexture(backBuffer_, backBufferImage);
-    return backBufferImage;
-}
-
-buw::Image4b IfcGeometryModelRenderer::captureImage()
-{
-    repaint();
-    return getBackBufferImage();
-}
-
-void IfcGeometryModelRenderer::setViewDirection(const buw::eViewDirection& direction, const bool fitViewToModel )
-{   
-	if( fitViewToModel )
-		this->fitViewToModel();
-    cameraController_->setViewDirection(buw::CameraController::getViewDirectionVector(direction));
-
-    // Operation takes 0.5 seconds
-    cameraController_->tick(1.0f);
-    camera_->tick(1.0f);
 }
